Rename module from api.services to pipeline
