+++ src/rust/rust-proto/src/lib.rs
@@ -9,14 +9,13 @@ pub(crate) mod graplinc {
                     ));
                 }
             }
-
-            pub(crate) mod services {
-                pub(crate) mod v1beta1 {
-                    include!(concat!(
-                        env!("OUT_DIR"),
-                        "/graplinc.grapl.api.services.v1beta1.rs"
-                    ));
-                }
+        }
+        pub(crate) mod pipeline {
+            pub(crate) mod v1beta1 {
+                include!(concat!(
+                    env!("OUT_DIR"),
+                    "/graplinc.grapl.api.services.v1beta1.rs"
+                ));
             }
         }
     }
@@ -25,4 +24,4 @@ pub(crate) mod graplinc {
 pub mod graph_descriptions;
 pub use graph_descriptions::node_property;
 
-pub mod services;
+pub mod pipeline;
